fix: split up components and resolve PR comments
