+++ pages/stands/[id].tsx
@@ -1,9 +1,5 @@
-import * as React from 'react';
-import { graphql } from 'react-relay';
-import { withDataAndLayout, WithDataAndLayoutProps } from '../../lib/withData';
-import styled from 'styled-components';
+import React from 'react';
 import { useRouter } from 'next/router';
-
 import * as _ from 'lodash';
 import Layout from '../../components/Layout';
